"build: generate modal template and functions"
