+++ css/style.css
@@ -18,6 +18,3 @@ h6,
 #add-new-button {
   cursor: cell;
 }
-.modal{
-  display: none;
-}
