+++ src/app.js
@@ -1,5 +1,5 @@
 import { renderApp, renderContent } from "./functions.js";
-import { createNewObject } from "./components/add-new-object/add.new.object.js";
+import { generateNewObject } from "./components/add-new-object/generate.new.object.js";
 
 document.addEventListener("DOMContentLoaded", () => {
   renderApp();
@@ -16,7 +16,7 @@ document.addEventListener("DOMContentLoaded", () => {
       document
         .getElementById("add-new-button")
         .addEventListener("click", () => {
-          createNewObject();
+          generateNewObject();
         });
     }
   });
