+++ src/components/add-new-object/generate.new.object.js
@@ -1,0 +1,29 @@
+import { generateModalFragment } from "../modal/modal.manager.js";
+import { addNewClass } from "../modal/templates-modal/add.class.modal.js";
+import { addNewTeacher } from "../modal/templates-modal/add.teacher.modal.js";
+import { addNewStudent } from "../modal/templates-modal/add.student.modal.js";
+
+export function generateNewObject() {
+  const pEl = document.getElementById("add-text");
+  const action = pEl.getAttribute("data-action");
+  switch (action) {
+    case "addNewClass":
+      generateModalFragment(addNewClass);
+      console.log(action);
+      break;
+    case "addNewTeacher":
+      generateModalFragment(addNewTeacher);
+      break;
+
+    case "addNewStudent":
+      generateModalFragment(addNewStudent);
+      break;
+
+    default:
+      break;
+  }
+  const dynamicModal = new bootstrap.Modal(
+    document.getElementById("dynamicModal")
+  );
+  dynamicModal.show();
+}
