+++ src/components/modal/templates-modal/add.student.modal.js
@@ -1,6 +1,8 @@
-export const addNewClass = `
+export const addNewStudent = document.createElement("div");
+
+addNewStudent.innerHTML = `
     <div class="modal-header">
-      <h5 class="modal-title">Add New Class</h5>
+      <h5 class="modal-title">Add New Student</h5>
       <button
         type="button"
         class="btn-close"
