+++ src/components/modal/templates-modal/main-modal.js
@@ -1,0 +1,4 @@
+const mainModal = `
+
+    
+
